Update Data::Alias to work with 5.12
The attached patch updates Data::Alias to work on Perl 5.12.
Specifically, in my tests it now passes tests on Perls
5.{8.{1..9},9.{0..5},10.{0..1},11.{0..5},12.{0..2},13.0}.  On 5.12.0+
the test suite gets a deprecation warning due to Switch being
deprecated in core, and on 5.13.1+ the test suite fails because Switch
has actually been removed from core.  As far as I know there's nothing
actually wrong with D:A in 5.13.1+, and it should work if something
else can be substituted for Switch (or those tests just skipped).

One of the tricky cases, alias with a braced block where the open
brace is on a new line, has a new failure mode in 5.11.2+.  It is
dependent on the reading in of the new line not triggering a
reallocation of the lexer buffer.  (Devel::Declare has the same
problem on all Perl versions and in more situations.)  This is
unavoidable: it is caused by D:A's hooks now being called earlier,
before the new line has been read in.  Only code with very long lines,
over about 4 KiB, should run into this problem.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>